add indicator to FAB
